refactor: Code moves in protocol updates mod
